Beta.

+++ app/src/main/java/com/ormanin/simpleweather/simpleweather/Model/WeatherModel/WeatherModel.java
@@ -17,7 +17,7 @@ public class WeatherModel {
     private String timezone;
     @SerializedName("offset")
     @Expose
-    private Integer offset;
+    private Double offset;
     @SerializedName("currently")
     @Expose
     private Currently currently;
@@ -70,11 +70,11 @@ public class WeatherModel {
         this.timezone = timezone;
     }
 
-    public Integer getOffset() {
+    public Double getOffset() {
         return offset;
     }
 
-    public void setOffset(Integer offset) {
+    public void setOffset(Double offset) {
         this.offset = offset;
     }
